Move `LocalFormLoader` and `FormStore` + other test helper classes into their own files
