+++ Lumiform/LumiformTests/Form Cache/LocalFormLoaderUseCaseTests.swift
@@ -6,6 +6,7 @@
 //
 
 import XCTest
+import Lumiform
 
 class LocalFormLoaderUseCaseTests: XCTestCase {
